+++ R/sig_fit_bootstrap_batch.R
@@ -2,7 +2,7 @@
 #'
 #' @inheritParams sig_fit
 #' @inheritParams sig_fit_bootstrap
-#' @param methods a subset of `c("LS", "QP", "SA")`.
+#' @param methods a subset of `c("NNLS", "QP", "SA")`.
 #' @param min_count minimal exposure in a sample, default is 1. Any patient has total exposure less
 #' than this value will be filtered out.
 #' @param p_val_thresholds a vector of relative exposure threshold for calculating p values.
@@ -51,7 +51,7 @@ sig_fit_bootstrap_batch <- function(catalogue_matrix,
   stopifnot(is.matrix(catalogue_matrix))
 
   set.seed(seed, kind = "L'Ecuyer-CMRG")
-  methods <- match.arg(methods, choices = c("LS", "QP", "SA"), several.ok = TRUE)
+  methods <- match.arg(methods, choices = c("NNLS", "QP", "SA"), several.ok = TRUE)
 
   timer <- Sys.time()
   send_info("Batch Bootstrap Signature Exposure Analysis Started.")